requirements.txt
python-dotenv==1.0.0
TTS==0.22.0
Flask-Caching==2.3.0
Flask-Limiter==3.3.0
Flask~=2.1.2
Jinja2>=3.1.2
Markdown
Pillow>=9.1.1
PyYAML>=6.0boto3
encodec
funcy
huggingface-hub>=0.14.1
scipy
tokenizers
torch==2.4.0
torchaudio==2.4.0
torchvision
tqdm
transformers
bandit
black
codecov
flake8
hypothesis>=6.14,<7
isort>=5.0.0,<6
jupyter
mypy
nbconvert
nbformat
pydocstyle
pylint
pytest
pytest-cov
joblib>=1.1.0
numba>=0.57.0
numpy==1.22.0
scipy==1.11.2
librosa>=0.10.0
llvmlite
fairseq==0.12.2
faiss-cpu==1.7.3
gradio==4.19.2
pydub>=0.25.1
soundfile>=0.12.1
ffmpeg-python>=0.2.0
tensorboardX
Jinja2>=3.1.2
json5
Markdown
matplotlib>=3.7.0
matplotlib-inline>=0.1.3
praat-parselmouth>=0.4.2
Pillow>=9.1.1
resampy>=0.4.2
scikit-learn
tensorboard
tornado>=6.1
werkzeug==2.1.2
uc-micro-py>=1.0.1
sympy>=1.11.1
tabulate>=0.8.10
PyYAML>=6.0
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
fsspec>=2022.11.0
absl-py>=1.2.0
audioread
uvicorn>=0.21.1
colorama>=0.4.5
pyworld==0.3.2
httpx>=0.24.1
torchcrepe==0.0.23
fastapi==0.109.1
pillow==10.2.0
ffmpeg
prefetch_generator
omegaconf==2.0.6
pytorch_lightning==1.2.10